refactor: optimize App component strcutrue
