+++ src/assets/App.less
@@ -98,7 +98,7 @@ body {
 
 @media screen and (max-width: 532px) {
   .top-area .home-bg {
-    background-image: url(./assets/image/home-bg-mobile.png);
+    background-image: url(../assets/image/home-bg-mobile.png);
     background-size: contain;
     background-repeat: no-repeat;
     background-position: center;
@@ -116,7 +116,7 @@ body {
 
 @media screen and (min-width: 532px) {
   .top-area .home-bg {
-    background-image: url(./assets/image/home-bg-pc.jpeg);
+    background-image: url(../assets/image/home-bg-pc.jpeg);
     background-size: contain;
     background-repeat: no-repeat;
     background-position: center;
