+++ src/main.tsx
@@ -3,8 +3,4 @@ import ReactDOM from 'react-dom/client'
 import App from './App'
 import 'antd/dist/reset.css'
 
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-)
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(<App />)
