+++ src/utils/constant.ts
@@ -1,16 +1,8 @@
-export const GitHubRepoAddress = 'https://github.com/33YANG/react_assignment'
+/** Github repo address */
+export const GITHUB_REPO_ADDRESS = 'https://github.com/33YANG/react_assignment'
 
-export const LanguageOptions = [
-  {
-    key: 'en',
-    label: 'English',
-  },
-  {
-    key: 'zh',
-    label: '中文',
-  },
-  {
-    key: 'jp',
-    label: '日本語',
-  },
-]
+/** Order Steps length, start with 0 */
+export const STEP_LENGTH = 3
+
+/** Max number of dishes */
+export const MAX_NUMBER_OF_DISHES = 10
